Create README.md
My design emphasizes the ideas of growth both in terms of the growth of crypto itself but also the idea that things that can grow from it. In the design, the Ethereum logo is broken into two parts representing a seed with a flower growing from half of it. Paying homage to crypto being based in cyberspace, the design is illustrated in a vaporware-esque style, featuring the git coin logo and the kanji for growth in computer windows.